code/webapp/requirements.txt
streamlit>=1.32
gspread>=6.0
google-auth>=2.27
pandas>=2.2
numpy>=1.26
openpyxl>=3.1
